# feat(core): optimize project locator perf