## Current Behavior
<!-- This is the behavior we have today -->

The `TargetProjectLocator` does a lot of extra work to find projects.

## Expected Behavior
<!-- This is the behavior we should expect with the changes in this PR -->

The `TargetProjectLocator` is more optimized.

## Results
On a large project, this function used to execute for ~15s, now it executes for ~1.1s.

Overall time to build the project graph used to take ~18s, now it takes ~4.4s. (75% improvement)

## Related Issue(s)
<!-- Please link the issue being fixed so it gets closed when this is merged. -->

Fixes #